Fix DAOS

+++ api/dev-test/src/main/java/com/vulcan/dev_test/domain/student/repository/StudentDao.java
@@ -4,13 +4,12 @@ import com.vulcan.dev_test.domain.student.Student;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 public interface StudentDao {
     Page<Student> list(String name, PageRequest pageRequest);
     Student store(Student student);
-    Optional<Student> show(BigInteger id);
+    Optional<Student> show(Long id);
     Student update(Student student);
     void destroy(Student student);
 }
